feat: allow entire `paragraph` inside of quotes
Lists and Quotes are now a subtype of detached modifiers with actually
allow an entire `paragraph` node (instead of just `paragraph_segment`)
to be placed inside of them.
This has the benefit of allowing `_line_break` which means we can avoid
an excessive use of the (soon to be introduced) ranged segments.