Remove markers
For now we will remove markers. They do not really serve any useful
purpose right now. One idea, in which they would (i.e. making them
document root nodes, which would allow combining multiple documents in a
single file) could be an interesting feature. It would even be
relatively easy to implement from the parser side, but its behavior
(especially around links) becomes very non-trivial.
Until we either come up with a solution to that, or an actually
real-life scenario where this feature would be useful (and not just
cool), we will remove markers.

The `%`-detached modifier will in the future be used for attributes, so
a different char for the marker would need to be found.