Part 2 - non-infectable slide behavior
This grammar generation is still quite the mess but I have not found a
better way yet... The parser size is now at 33MB but we have quite some
odd cases here. Especially problematic is the nesting levels being
coupled between the parent object of the slide and its contents.

E.g.:

```norg
-- :
   > this does not work because the outer list item is level 2, so the
     inner quote item also needs to be at least that level
```

Also:

```norg
> :
  --- this also does not work because allowing arbitrary nesting levels
      in here results in *absolutely frickin massive* parsers 😱
```

I think we need to figure out a better way on how to deal with
infectability if we want to implement this properly.